Chore: move posts page to root dir
